+++ tests/test_classification.py
@@ -31,3 +31,14 @@ class TestClassification:
         expected_result = SLIX.io.imread('./tests/files/cluster/results/cluster_inclination_mask.tiff')
         assert numpy.all(cluster == expected_result)
 
+    def test_full_mask(self):
+        max_image = SLIX.io.imread('./tests/files/cluster/cluster_max.tiff')
+        high_prominence_peaks = SLIX.io.imread('./tests/files/cluster/cluster_high_prominence_peaks.tiff')
+        low_prominence_peaks = SLIX.io.imread('./tests/files/cluster/cluster_low_prominence_peaks.tiff')
+        peakdistance = SLIX.io.imread('./tests/files/cluster/cluster_peakdistance.tiff')
+        cluster = SLIX.classification.full_mask(high_prominence_peaks, low_prominence_peaks,
+                                                peakdistance, max_image)
+
+        expected_result = SLIX.io.imread('./tests/files/cluster/results/cluster_classification_mask.tiff')
+        assert numpy.all(cluster == expected_result)
+